Provide a way to disable sharpSAT tests when building it
With these changes, both `BUILD_TESTING` and `SHARPSAT_BUILD_TESTING`
need to be enabled for the tests to be built and run, or either can
be turned off to disable tests. `BUILD_TESTING` is a standard CMake
variable supported by CTest, `SHARPSAT_BUILD_TESTING` is a custom
option that allows more fine-grained control over the project.